Renaming ARM to AArch64 for more accuracy
