+++ kernel/AArch64/Boot/ExceptionLevel.h
@@ -1,5 +1,5 @@
-#ifndef KERNEL_ARM_BOOT_EXCEPTIONLEVEL_H
-#define KERNEL_ARM_BOOT_EXCEPTIONLEVEL_H
+#ifndef KERNEL_AARCH64_BOOT_EXCEPTIONLEVEL_H
+#define KERNEL_AARCH64_BOOT_EXCEPTIONLEVEL_H
 
 namespace AArch64
 {
@@ -12,4 +12,4 @@ namespace AArch64
     }
 }
 
-#endif // KERNEL_ARM_BOOT_EXCEPTIONLEVEL_H
+#endif // KERNEL_AARCH64_BOOT_EXCEPTIONLEVEL_H
