+++ kernel/AArch64/Boot/MMU.h
@@ -1,5 +1,5 @@
-#ifndef KERNEL_ARM_BOOT_MMU_H
-#define KERNEL_ARM_BOOT_MMU_H
+#ifndef KERNEL_AARCH64_BOOT_MMU_H
+#define KERNEL_AARCH64_BOOT_MMU_H
 
 namespace AArch64
 {
@@ -19,4 +19,4 @@ namespace AArch64
     }
 }
 
-#endif // KERNEL_ARM_BOOT_MMU_H
+#endif // KERNEL_AARCH64_BOOT_MMU_H
