+++ kernel/AArch64/RegisterDefines.h
@@ -1,7 +1,7 @@
 // Included from assembly, so can't have anything fancy in here
 
-#ifndef KERNEL_ARM_REGISTER_DEFINES_H
-#define KERNEL_ARM_REGISTER_DEFINES_H
+#ifndef KERNEL_AARCH64_REGISTER_DEFINES_H
+#define KERNEL_AARCH64_REGISTER_DEFINES_H
 
 ///////////////////////////////////////////////////////////////////////////////
 // SCTLR_EL1 register flags - See D10.2.100 in the ARMv8 manual
@@ -107,4 +107,4 @@
 #define ESR_ELx_EC_SVC64    0x15    // 0b010101 - exception caused by SVC instruction in AArch64 state
 #define ESR_ELx_EC_DABT_LOW 0x24    // 0b100100 - data abort from a lower exception level
 
-#endif // KERNEL_ARM_REGISTER_DEFINES_H
+#endif // KERNEL_AARCH64_REGISTER_DEFINES_H
