+++ kernel/AArch64/SchedulerDefines.h
@@ -1,0 +1,8 @@
+// Included from assembly, so can't have anything fancy in here
+
+#ifndef KERNEL_AARCH64_SCHEDULER_DEFINES_H
+#define KERNEL_AARCH64_SCHEDULER_DEFINES_H
+
+#define TASK_STRUCT_CONTEXT_OFFSET 0
+
+#endif // KERNEL_AARCH64_SCHEDULER_DEFINES_H
