--- kernel/ARM/SchedulerDefines.h
@@ -1,8 +1,0 @@
-// Included from assembly, so can't have anything fancy in here
-
-#ifndef KERNEL_ARM_SCHEDULER_DEFINES_H
-#define KERNEL_ARM_SCHEDULER_DEFINES_H
-
-#define TASK_STRUCT_CONTEXT_OFFSET 0
-
-#endif // KERNEL_ARM_SCHEDULER_DEFINES_H
