+++ kernel/Main.cpp
@@ -1,4 +1,4 @@
-#include "ARM/MMUDefines.h"
+#include "AArch64/MMUDefines.h"
 #include "Peripherals/DeviceTree.h"
 #include "ExceptionVectorHandlers.h"
 #include "IRQ.h"
