+++ kernel/Scheduler.S
@@ -1,4 +1,4 @@
-#include "ARM/SchedulerDefines.h"
+#include "AArch64/SchedulerDefines.h"
 
 .globl cpu_switch_to
 cpu_switch_to:
